Implement Phase 2 PixelTracks CA Extension
- Enable OT SoA producers, increase SoA size for pixelTracks.
- Update CAHitNtuplet to support Phase2 OT CA extension.
- Update also the configuration parameters to include the 3 additional
barrel layers from the OT.
- Correctly compute the index used to address rotations.
The index in the SiStipHitSoA to address the proper rotation of the
module is now correctly computed by couting all the preceeding pixel
modules and simply incrementing that number by one, for every new
P-module in the OT barrel.
The code and logic is extremely fragile and relies on a couple of
assumptions:
- all pixel modules will be used in the CA, not matter what.
- the order of modules in Phase2OTRecHitsSoAConverter and in
CAHitNtuplet is the same.
- Add producer to merge Pixel and tracker recHits
- Modify PixelTrack converter (SoA -> legacy) to be able to handle the extension to OT in Phase-2
- Update hard-coded constants to be able to run CA+Extension.
- Add new TrackerTraits topology
- Fix CA to work both with and without OT extension
The traits guarding Phase2 and Phase2OT should now be guarded correctly

Co-Authored-By: Adriano Di Florio <adriano.di.florio@cern.ch>
Co-Authored-By: Marco Musich <marco.musich@cern.ch>
Co-Authored-By: Jan Schulz <jan.gerrit.schulz@cern.ch>
Co-Authored-By: Luca Ferragina <ferragina.luca@gmail.com>